More stdlib refactoring
