+++ arlunio/lib/mask/mask.py
@@ -21,36 +21,6 @@ def MaskAdd(
         The first mask
     b:
         The second mask
-
-    Examples
-    --------
-    This definition is used implicitly when adding two masks together
-
-    .. arlunio-image::
-       :include-code: before
-
-       import arlunio as ar
-       from arlunio.lib import Circle, Square
-
-       c1 = Square(xc=-0.25, yc=-0.25, size=0.55)
-       c2 = Circle(xc=0.25, yc=0.25, r=0.7)
-
-       c = c1 + c2
-       image = ar.fill(c(width=1920, height=1080))
-
-    Or can be used directly
-
-    .. arlunio-image::
-       :include-code: before
-
-       import arlunio as ar
-       from arlunio.lib import Circle, MaskAdd, Square
-
-       a = Square(xc=-0.25, yc=-0.25, size=0.55)
-       b = Circle(xc=0.25, yc=0.25, r=0.7)
-
-       c = MaskAdd(a=a, b=b)
-       image = ar.fill(c(width=1920, height=1080))
     """
 
     return ar.any(a(width=width, height=height), b(width=width, height=height))
@@ -76,36 +46,6 @@ def MaskSub(
         The first "base" mask
     b:
         The second mask that defines the region to remove from :code:`a`
-
-    Examples
-    --------
-    This definition is used implicitly when subtracting one mask from another
-
-    .. arlunio-image::
-       :include-code: before
-
-       import arlunio as ar
-       from arlunio.lib import Circle, Square
-
-       c1 = Square(xc=-0.25, yc=-0.25, size=0.55)
-       c2 = Circle(xc=0.25, yc=0.25, r=0.7)
-
-       c = c1 - c2
-       image = ar.fill(c(width=1920, height=1080))
-
-    Or can be used directly
-
-    .. arlunio-image::
-       :include-code: before
-
-       import arlunio as ar
-       from arlunio.lib import Circle, MaskSub, Square
-
-       a = Square(xc=-0.25, yc=-0.25, size=0.55)
-       b = Circle(xc=0.25, yc=0.25, r=0.7)
-
-       c = MaskSub(a=b, b=a)
-       image = ar.fill(c(width=1920, height=1080))
     """
     return ar.all(
         a(width=width, height=height), ar.invert(b(width=width, height=height))
@@ -127,35 +67,5 @@ def MaskMul(
         The first mask
     b:
         The second mask
-
-    Examples
-    --------
-    This definition is used implicitly when multiplying two masks together
-
-    .. arlunio-image::
-       :include-code: before
-
-       import arlunio as ar
-       from arlunio.lib import Circle, Square
-
-       c1 = Square(xc=-0.25, yc=-0.25, size=0.55)
-       c2 = Circle(xc=0.25, yc=0.25, r=0.7)
-
-       c = c1 * c2
-       image = ar.fill(c(width=1920, height=1080))
-
-    Or can be used directly
-
-    .. arlunio-image::
-       :include-code: before
-
-       import arlunio as ar
-       from arlunio.lib import Circle, MaskMul, Square
-
-       a = Square(xc=-0.25, yc=-0.25, size=0.55)
-       b = Circle(xc=0.25, yc=0.25, r=0.7)
-
-       c = MaskMul(a=a, b=b)
-       image = ar.fill(c(width=1920, height=1080))
     """
     return ar.all(a(width=width, height=height), b(width=width, height=height))
